fix(admin): Add admin navigation and access gate
