+++ tavern/internal/www/src/components/admin-access-gate/index.tsx
@@ -1,0 +1,1 @@
+export { AdminAccessGate } from "./AdminAccessGate";
