--- tavern/internal/www/src/components/page-wrapper/MobileNav.tsx
@@ -1,123 +1,0 @@
-import { Fragment } from 'react'
-import { Dialog, Transition } from '@headlessui/react'
-import {
-    Bars3Icon,
-    XMarkIcon,
-} from '@heroicons/react/24/outline'
-import { Link } from 'react-router-dom';
-import { NavigationItemType } from '../../utils/consts';
-import { classNames } from '../../utils/utils';
-
-type MobileNavProps = {
-    sidebarOpen: boolean;
-    handleSidebarOpen: (arg: boolean) => void;
-    currNavItem?: string;
-    navigation: Array<NavigationItemType>
-}
-
-const MobileNav = ({ handleSidebarOpen, sidebarOpen, currNavItem, navigation }: MobileNavProps) => {
-    return (
-        <>
-            <div className="sticky top-0 z-40 flex items-center gap-x-6 bg-gray-900 py-4 px-4 shadow-sm sm:px-6 lg:hidden">
-                <button type="button" className="-m-2.5 p-2.5 text-gray-400 lg:hidden" onClick={() => handleSidebarOpen(true)}>
-                    <span className="sr-only">Open sidebar</span>
-                    <Bars3Icon className="h-6 w-6" aria-hidden="true" />
-                </button>
-                <div className="flex-1 text-sm font-semibold leading-6 text-white">Realm</div>
-                <div className='text-sm font-semibold leading-6 text-white'>Logout</div>
-            </div>
-            <Transition.Root show={sidebarOpen} as={Fragment}>
-                <Dialog as="div" className="relative z-50 lg:hidden" onClose={handleSidebarOpen}>
-                    <Transition.Child
-                        as={Fragment}
-                        enter="transition-opacity ease-linear duration-300"
-                        enterFrom="opacity-0"
-                        enterTo="opacity-100"
-                        leave="transition-opacity ease-linear duration-300"
-                        leaveFrom="opacity-100"
-                        leaveTo="opacity-0"
-                    >
-                        <div className="fixed inset-0 bg-gray-900/80" />
-                    </Transition.Child>
-
-                    <div className="fixed inset-0 flex">
-                        <Transition.Child
-                            as={Fragment}
-                            enter="transition ease-in-out duration-300 transform"
-                            enterFrom="-translate-x-full"
-                            enterTo="translate-x-0"
-                            leave="transition ease-in-out duration-300 transform"
-                            leaveFrom="translate-x-0"
-                            leaveTo="-translate-x-full"
-                        >
-                            <Dialog.Panel className="relative mr-16 flex w-full max-w-xs flex-1">
-                                <Transition.Child
-                                    as={Fragment}
-                                    enter="ease-in-out duration-300"
-                                    enterFrom="opacity-0"
-                                    enterTo="opacity-100"
-                                    leave="ease-in-out duration-300"
-                                    leaveFrom="opacity-100"
-                                    leaveTo="opacity-0"
-                                >
-                                    <div className="absolute top-0 left-full flex w-16 justify-center pt-5">
-                                        <button type="button" className="-m-2.5 p-2.5" onClick={() => handleSidebarOpen(false)}>
-                                            <span className="sr-only">Close sidebar</span>
-                                            <XMarkIcon className="h-6 w-6 text-white" aria-hidden="true" />
-                                        </button>
-                                    </div>
-                                </Transition.Child>
-                                {/* Sidebar component*/}
-                                <div className="flex grow flex-col gap-y-5 overflow-y-auto bg-gray-900 px-6 pb-2 ring-1 ring-white/10">
-                                    <div className="flex h-16 shrink-0 items-center">
-                                    </div>
-                                    <nav className="flex flex-1 flex-col">
-                                        <ul className="flex flex-1 flex-col gap-y-7">
-                                            <li>
-                                                <ul className="-mx-2 space-y-1">
-                                                    {navigation.map((item) => (
-                                                        <li key={item.name}>
-                                                            {item.internal ? (
-                                                                <Link
-                                                                    to={item.href}
-                                                                    className={classNames(
-                                                                        item.name === currNavItem
-                                                                            ? 'bg-gray-800 text-white'
-                                                                            : 'text-gray-400 hover:text-white hover:bg-gray-800',
-                                                                        'group flex gap-x-3 rounded-md p-2 text-sm leading-6 font-semibold'
-                                                                    )}
-                                                                >
-                                                                    <item.icon className="h-6 w-6 shrink-0" aria-hidden="true" />
-                                                                    {item.name}
-                                                                </Link>
-                                                            ) : (
-                                                                <a
-                                                                    href={item.href}
-                                                                    target={item?.target ? '__blank' : undefined}
-                                                                    className={classNames(
-                                                                        item.name === currNavItem
-                                                                            ? 'bg-gray-800 text-white'
-                                                                            : 'text-gray-400 hover:text-white hover:bg-gray-800',
-                                                                        'group flex gap-x-3 rounded-md p-2 text-sm leading-6 font-semibold'
-                                                                    )}
-                                                                >
-                                                                    <item.icon className="h-6 w-6 shrink-0" aria-hidden="true" />
-                                                                    {item.name}
-                                                                </a>
-                                                            )}
-                                                        </li>
-                                                    ))}
-                                                </ul>
-                                            </li>
-                                        </ul>
-                                    </nav>
-                                </div>
-                            </Dialog.Panel>
-                        </Transition.Child>
-                    </div>
-                </Dialog>
-            </Transition.Root>
-        </>
-    );
-}
-export default MobileNav;
