--- tavern/internal/www/src/components/page-wrapper/index.tsx
@@ -1,1 +1,0 @@
-export {PageWrapper} from "./PageWrapper";
