+++ tavern/internal/www/src/features/page-wrapper/index.tsx
@@ -1,0 +1,1 @@
+export {PageWrapper} from "./PageWrapper";
